pyproject.toml
[tool.poetry]
name = "line-running-consultation"
version = "0.1.0"
description = "陸上競技練習相談LINEアプリケーション"
authors = ["Development Team"]
packages = [{include = "app"}, {include = "api"}, {include = "core"}, {include = "infrastructure"}, {include = "shared"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
redis = "^5.0.0"
alembic = "^1.13.0"
line-bot-sdk = "^3.8.0"
google-generativeai = "^0.3.0"
langchain = "^0.1.0"
langchain-google-genai = "^0.0.8"
langchain-community = "^0.0.13"
chromadb = "^0.4.0"
sentence-transformers = "^2.2.0"
python-dotenv = "^1.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "^4.1.0"
python-multipart = "^0.0.6"
aiohttp = "^3.9.0"
pandas = "^2.1.0"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"
factory-boy = "^3.3.0"
coverage = "^7.3.0"
black = "^23.9.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 